onebase/requirements.txt
typer
rich
pyyaml
pydantic
# langchain 相关依赖
langchain
langchain-postgres
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-google-vertexai
langchain-deepseek
langchain-ollama
langchain-modelscope-integration
langchain-groq
qianfan
dashscope
zhipuai
psycopg[binary]
python-dotenv
pypdf